Move web/api related modules to 'Web' folder.
